Move tests for Java classes to `memeid`
