+++ memeid4s-http4s/src/test/scala/memeid4s/http4s/InstancesSpec.scala
@@ -19,7 +19,7 @@ package memeid4s.http4s
 import cats.syntax.show._
 
 import memeid4s.UUID
-import memeid4s.arbitrary.instances._
+import memeid.arbitrary.instances._
 import memeid4s.cats.instances._
 import memeid4s.http4s.instances._
 import org.http4s.dsl.impl.QueryParamDecoderMatcher
